feat: 메인 탭 통합
